Changes to move from od to od4.4.
git-svn-id: https://svn.opendtect.org/od/branches/od4.4@26214 a985855d-d384-4b45-ae33-31a410998c3c

+++ src/Algo/reflectivitysampler.cc
@@ -102,7 +102,7 @@ bool ReflectivitySampler::doWork( od_int64 start, od_int64 stop, int threadidx )
 	const float anglesampling = -time * df;
 	while ( ptr!=stopptr )
 	{
-	    const float angle = (float) ( 2*M_PI *anglesampling * freqidx );
+	    const float angle = 2*M_PI *anglesampling * freqidx;
 	    const float freq = df * freqidx;
 	    const float_complex cexp = float_complex( cos(angle), sin(angle) );
 	    const float_complex cpexref = cexp * reflectivity;
